feat: AddMetadataTransform rename
