+++ src/main/java/net/tk/kafka/connect/transforms/AddMetadataConfig.java
@@ -1,13 +1,13 @@
-package net.tk.transformer;
+package net.tk.kafka.connect.transforms;
 
 import org.apache.kafka.common.config.AbstractConfig;
 import org.apache.kafka.common.config.ConfigDef;
 
 import java.util.Map;
 
-public class SchemaWrappingConfig extends AbstractConfig {
+public class AddMetadataConfig extends AbstractConfig {
 
-    public SchemaWrappingConfig(ConfigDef configDef, Map<?, ?> originals) {
+    public AddMetadataConfig(ConfigDef configDef, Map<?, ?> originals) {
         super(configDef, originals, false);
     }
